fix(blog): keep sitemap lastmod importable without blog-site node_modules
main is red. #7646 stopped postinstall installing blog-site, but
tests/blog-seo-contract.test.mjs reaches the lastmod logic through
`await import('../blog-site/astro.config.mjs')`, and that config's first
line is `import { defineConfig } from 'astro/config'`. With the workspace
no longer installed on the unit job, that resolves to
ERR_MODULE_NOT_FOUND: Cannot find package 'astro'.

This is my miss on #7646. I enumerated the blog-site consumers with a
truncated `grep -rln ... | head -15` and concluded no test needed the
workspace's node_modules. The real list is 19 files, and this was one of
the four outside that window.

buildPostDateMap and its helpers use only node: builtins. Nothing about
computing sitemap dates needs Astro; only the config's own framework
imports do. Move that block to blog-site/post-dates.mjs, have
astro.config.mjs consume it, and point the test at the pure module.

astro.config.mjs deliberately does not re-export buildPostDateMap. It was
exported only for that test, nothing else imports the config, and leaving
the re-export in place would let the same bug back in through the same
door.

The separation is right on its own terms. Sitemap lastmod derivation is
not framework configuration, and it lived in astro.config.mjs only
because that is where the sitemap integration reads it.

Reproduced and verified with blog-site/node_modules deleted, which is the
condition the unit job now runs in. On main's files the test fails with
Cannot find package 'astro'. With this change the same command passes.
All 19 blog-site-referencing suites pass in that state (434 tests), and
`npm run build:blog:raw` still installs and builds 76 pages with a
sitemap through the restructured config.
